+++ gm2m/monkeypatch.py
@@ -103,13 +103,13 @@ try:
                 ', '.join([
                     "id",
                     gfbn_new(old_names['src'])[0].column,
-                    gfbn_new(old_names['tgt_pk'])[0].column,
+                    gfbn_new(old_names['tgt_fk'])[0].column,
                     gfbn_new(old_names['tgt_ct'])[0].column,
                 ]),
                 ', '.join([
                     "id",
                     gfbn_old(old_names['src'])[0].column,
-                    gfbn_old(old_names['tgt_pk'])[0].column,
+                    gfbn_old(old_names['tgt_fk'])[0].column,
                     gfbn_old(old_names['tgt_ct'])[0].column,
                 ]),
                 self.quote_name(old_field.rel.through._meta.db_table),
@@ -120,6 +120,24 @@ try:
             return _alter_many_to_many_sqlite0(self, model, old_field,
                                                new_field, strict)
     DatabaseSchemaEditor._alter_many_to_many = _alter_many_to_many
+
+    from  django.db.migrations.autodetector import MigrationAutodetector
+
+    def only_relation_agnostic_fields(self, fields):
+        """
+        We only change the way the 'to' key is deleted from the dict
+        (as GM2MField.deconstruct does not return a 'to' kwarg)
+        """
+        fields_def = []
+        for __, field in fields:
+            deconstruction = self.deep_deconstruct(field)
+            if field.rel and field.rel.to:
+                deconstruction[2].pop('to', None)
+            fields_def.append(deconstruction)
+        return fields_def
+
+    MigrationAutodetector.only_relation_agnostic_fields = \
+        only_relation_agnostic_fields
 
 
 except (ImportError, AttributeError):